review comments, refactor config
