+++ zio-http/jvm/src/main/scala/zio/http/netty/client/ClientSSLConverter.scala
@@ -18,18 +18,16 @@ package zio.http.netty.client
 
 import java.io.{File, FileInputStream, InputStream}
 import java.security.KeyStore
-import javax.net.ssl.TrustManagerFactory
 
+import javax.net.ssl.TrustManagerFactory
 import scala.util.Using
 
 import zio.Config.Secret
 import zio.stacktracer.TracingImplicits.disableAutoTrace
-
 import zio.http.ClientSSLConfig
-import zio.http.ClientSSLConfig.{FromClientCertFile, FromClientCertResource}
-
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory
 import io.netty.handler.ssl.{SslContext, SslContextBuilder}
+import zio.http.ClientSSLCertConfig.{FromClientCertFile, FromClientCertResource}
 object ClientSSLConverter {
   private def trustStoreToSslContext(
     trustStoreStream: InputStream,
